Split cloud tests into modules
