+++ test-cloud/test-cloud-optaweb/src/test/java/org/kie/cloud/optaweb/OptawebEmployeeRosteringIntegrationTest.java
@@ -1,4 +1,19 @@
-package org.kie.cloud.integrationtests.optaweb;
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.cloud.optaweb;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -7,8 +22,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.kie.cloud.api.DeploymentScenarioBuilderFactory;
 import org.kie.cloud.api.scenario.EmployeeRosteringScenario;
-import org.kie.cloud.integrationtests.AbstractMethodIsolatedCloudIntegrationTest;
-import org.kie.cloud.integrationtests.testproviders.OptawebEmployeeRosteringTestProvider;
+import org.kie.cloud.optaweb.testprovider.OptawebEmployeeRosteringTestProvider;
+import org.kie.cloud.tests.common.AbstractMethodIsolatedCloudIntegrationTest;
 
 public class OptawebEmployeeRosteringIntegrationTest extends AbstractMethodIsolatedCloudIntegrationTest<EmployeeRosteringScenario> {
