+++ test-cloud/test-cloud-optaweb/src/test/java/org/kie/cloud/optaweb/testprovider/OptawebEmployeeRosteringTestProvider.java
@@ -1,4 +1,19 @@
-package org.kie.cloud.integrationtests.testproviders;
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.cloud.optaweb.testprovider;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -11,7 +26,7 @@ import java.util.Set;
 import org.assertj.core.api.Assertions;
 import org.jboss.resteasy.client.jaxrs.ResteasyClient;
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
-import org.kie.cloud.integrationtests.optaweb.rest.OptaWebObjectMapperResolver;
+import org.kie.cloud.optaweb.rest.OptaWebObjectMapperResolver;
 import org.optaweb.employeerostering.restclient.ServiceClientFactory;
 import org.optaweb.employeerostering.shared.employee.Employee;
 import org.optaweb.employeerostering.shared.employee.EmployeeAvailabilityState;
